% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climatologia.R
\name{station}
\alias{station}
\title{CLIMATOLOGIA}
\usage{
station(est, sheet)
}
\arguments{
\item{est}{: es el nombre de la estacion (ej: '2_Abapo.xlsx')}

\item{sheet}{: es un numero entero (cada hoja es una varible meteorologica)}
}
\description{
Este es un algoritmo para las estaciones de SENAMHI-BOLIVIA, el fin de este trabajo es para
tener un mejor manejo con los datos que monitorea las estaciones que estan en el dominio de Bolivia.
}
\details{
Originalmente los datos vienen en una extension .xlsx, una extension que sismet tiene almacenado los
datos monitoreados por las estaciones senamhi-Bolivia; el algoritmo simplifica considerablemente el proceso de transformar
y ordenar los datos a una extension mas frecuente por la comunidad de analistas de datos.

 Este trabajo fue posible gracias a mis compa\~neros de climatologia de la carrera de fisica 1/2021.
}
\author{
Nihel Ruiz Atanacio
}
